Create README.md
This is a vision enhancer based module specifically for the BLIND VICTIMS. The system is designed in such a way in which the blind person can take the help of AN APPLICATION which inturn sends Real Time Frames to the LAPTOP-BASED WIRELESS NETWORKED SYSTEM. It works on REAL-TIME OBJECT DETECTION using SSD_MOBILENET algorithm and TENSORFLOW APIs . It has a core feature of approximate distance calculation and Voice - Based wireless feedack generation w.r.t the DISTANCE CALCULATION. It makes the work of Blind easy,efficient and reliable by sending wireless Voice based feedback whether the particular object is either too close to him or is it at a safer distance. The same system can be used from Obstacle Detection.